refactor: move Utils to main
