--- bls/src/main/java/org/dashj/bls/Utils.java
@@ -1,7 +1,0 @@
-package org.dashj.bls;
-
-import com.google.common.io.BaseEncoding;
-
-public class Utils {
-    public static final BaseEncoding HEX = BaseEncoding.base16().lowerCase();
-}
